@@ -2,14 +2,29 @@
 	import { base } from '$app/paths';
 	import Icon from './Icon.svelte';
 
-	export let link: string;
-	export let details = '';
-	export let detailsLang: 'de' | 'en' = 'de';
-	export let invertedStyle = false;
-	export let secondaryStyle = false;
-	export let download = false;
-	export let allowReferrer = false;
-	export let noBase = false;
+	interface Props {
+		link: string;
+		details?: string;
+		detailsLang?: 'de' | 'en';
+		invertedStyle?: boolean;
+		secondaryStyle?: boolean;
+		download?: boolean;
+		allowReferrer?: boolean;
+		noBase?: boolean;
+		children?: import('svelte').Snippet;
+	}
+
+	let {
+		link,
+		details = '',
+		detailsLang = 'de',
+		invertedStyle = false,
+		secondaryStyle = false,
+		download = false,
+		allowReferrer = false,
+		noBase = false,
+		children
+	}: Props = $props();
 </script>
 
 {#if !download}
@@ -20,16 +35,16 @@
 		target="_blank"
 		rel={allowReferrer ? 'noopener no-referrer-when-downgrade' : 'noopener noreferrer'}
 	>
-		<span aria-hidden="true"><slot /></span>
+		<span aria-hidden="true">{@render children?.()}</span>
 		<span class="sr-only"
-			><slot /><span lang={detailsLang}>{details}</span> (öffnet neues Fenster)</span
+			>{@render children?.()}<span lang={detailsLang}>{details}</span> (öffnet neues Fenster)</span
 		>
 		<Icon svg="open-in-new-tab" size="20" color={invertedStyle ? 'white' : 'blue'} />
 	</a>
 {:else}
 	<a href={!noBase ? base + '/' + link : link} download class:invertedStyle class:secondaryStyle>
-		<span aria-hidden="true"><slot /></span>
-		<span class="sr-only"><slot /><span lang={detailsLang}>{details}</span></span>
+		<span aria-hidden="true">{@render children?.()}</span>
+		<span class="sr-only">{@render children?.()}<span lang={detailsLang}>{details}</span></span>
 		<Icon svg="download" size="20" color={invertedStyle ? 'white' : 'blue'} />
 	</a>
 {/if}
