@@ -1,9 +1,13 @@
 <script lang="ts">
 	import { base } from '$app/paths';
 	import { allowYouTubeEmbed } from '$lib/store';
-	export let src: string = '';
-	export let linkToYouTube: string = '';
-	export let videoTitle: string = '';
+	interface Props {
+		src?: string;
+		linkToYouTube?: string;
+		videoTitle?: string;
+	}
+
+	let { src = '', linkToYouTube = '', videoTitle = '' }: Props = $props();
 
 	const randomId =
 		Math.random().toString(36).substring(2, 10) + Math.random().toString(36).substring(2, 10);
