@@ -4,9 +4,11 @@
 	import { onMount } from 'svelte';
 	import Icon from '../Icon.svelte';
 
-	let menu: HTMLDialogElement;
+	let menu: HTMLDialogElement | undefined = $state();
 
-	export function toggleDisplay() {
+	export function toggleDisplay(e: Event) {
+		e.stopPropagation();
+
 		if (menu?.open) {
 			menu?.close();
 			document?.body.removeAttribute('style');
@@ -19,13 +21,15 @@
 
 	function handleBackdropClick(e: MouseEvent) {
 		const dialogDimensions = menu?.getBoundingClientRect();
+		if (!dialogDimensions) return;
+
 		if (
 			e.clientX < dialogDimensions.left ||
 			e.clientX > dialogDimensions.right ||
 			e.clientY < dialogDimensions.top ||
 			e.clientY > dialogDimensions.bottom
 		) {
-			toggleDisplay();
+			toggleDisplay(e);
 		}
 	}
 
@@ -45,7 +49,7 @@
 			});
 		});
 
-		menuAttrObserver.observe(menu, {
+		menuAttrObserver.observe(menu!, {
 			attributes: true
 		});
 
@@ -54,25 +58,21 @@
 		});
 	});
 
-	function goToSite(site: string) {
-		toggleDisplay();
+	function goToSite(site: string, e: Event) {
+		e.preventDefault();
+		toggleDisplay(e);
 		goto(base + site);
 	}
 </script>
 
-<!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
-<!-- svelte-ignore a11y-click-events-have-key-events -->
-<dialog
-	inert
-	bind:this={menu}
-	aria-label="Hauptmenü"
-	on:click|stopPropagation={(e) => handleBackdropClick(e)}
->
+<!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<dialog inert bind:this={menu} aria-label="Hauptmenü" onclick={(e) => handleBackdropClick(e)}>
 	<div class="menu-header">
 		<button
 			autofocus
 			type="button"
-			on:click|stopPropagation={() => toggleDisplay()}
+			onclick={(e) => toggleDisplay(e)}
 			aria-label="Hauptmenü schließen"
 		>
 			<Icon svg="close" size="parent" />
@@ -82,40 +82,27 @@
 	<div class="content-wrapper">
 		<ul>
 			<li>
-				<a
-					on:click|preventDefault|stopPropagation={() => goToSite('/#projektergebnisse')}
-					href="{base}/#projektergebnisse"
-				>
+				<a onclick={(e) => goToSite('/#projektergebnisse', e)} href="{base}/#projektergebnisse">
 					Projektergebnisse
 				</a>
 			</li>
 			<li>
-				<a
-					on:click|preventDefault|stopPropagation={() => goToSite('/webinaraufnahmen')}
-					href="{base}/webinaraufnahmen">Webinaraufnahmen</a
+				<a onclick={(e) => goToSite('/webinaraufnahmen', e)} href="{base}/webinaraufnahmen"
+					>Webinaraufnahmen</a
 				>
 			</li>
 			<li>
-				<a on:click|preventDefault|stopPropagation={() => goToSite('/team')} href="{base}/team"
-					>Team</a
-				>
+				<a onclick={(e) => goToSite('/team', e)} href="{base}/team">Team</a>
 			</li>
 			<li>
-				<a
-					on:click|preventDefault|stopPropagation={() => goToSite('/aktuelles')}
-					href="{base}/aktuelles">Aktuelles</a
-				>
+				<a onclick={(e) => goToSite('/aktuelles', e)} href="{base}/aktuelles">Aktuelles</a>
 			</li>
 			<li>
-				<a
-					on:click|preventDefault|stopPropagation={() => goToSite('/kontakt')}
-					href="{base}/kontakt">Kontakt</a
-				>
+				<a onclick={(e) => goToSite('/kontakt', e)} href="{base}/kontakt">Kontakt</a>
 			</li>
 			<li>
-				<a
-					on:click|preventDefault|stopPropagation={() => goToSite('/publikationen')}
-					href="{base}/publikationen">Publikationen</a
+				<a onclick={(e) => goToSite('/publikationen', e)} href="{base}/publikationen"
+					>Publikationen</a
 				>
 			</li>
 		</ul>
